Implement min_appearance token filter OOV token
The min_appearance filter ensures that no tokens are included in the
tokenizer.json file unless they appear at least X times. In this commit,
X is 3. The OOV token is assigned to a specific value if it doesn't
exist in the tokenizer. In this commit the OOV token is 0.